+++ CHANGELOG.md
@@ -19,7 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Playwright test to filter search results with dropdowns (DR-3659)
 - Added GitHub Actions workflow to run Playwright tests on PRs (DR-3721)
 - Added Playwright tests to clear search results filters (DR-3659)
-- Added Playwright test for the division page (DR-3655)
+- Added Playwright  test where a user clicks Divisions from home page navigation menu (DR-3655)
+- Added Playwright test to verify navigation menu on Divisons page (DR-3655)
+- Added Playwright test to verify Division's name headings and the see more links for each division (DR-3655)
 
 ## [0.4.2] 2025-05-22